refacto drage and drop list
