+++ packages/@coorpacademy-components/src/molecule/draggable/test/fixtures/default.js
@@ -1,0 +1,11 @@
+export default {
+  props: {
+    id: 'draggableId',
+    dragging: true,
+    onDragStart: () => console.log('onDragStart'),
+    onDragOver: () => console.log('onDragOver'),
+    onDragLeave: () => console.log('onDragLeave'),
+    onDrop: () => console.log('onDrop'),
+    children: []
+  }
+};
