+++ packages/@coorpacademy-components/src/molecule/setup-section/style.css
@@ -12,7 +12,7 @@
   display: flex;
   align-items: center;
   justify-content: space-between;
-  padding: 0 24px;
+  padding: 0 24px 0 48px;
   box-sizing: border-box;
 }
 
@@ -37,32 +37,7 @@
   margin-left: 30px;
   margin-right: 16px;
 }
-.arrowSection {
-  display: none
-}
-.arrow {
-  left: 50%;
-  top: 50%;
-  margin: 0 0 0 24px;
-  width: 30px;
-  height: 30px;
-  cursor: pointer;
-}
-
-.arrow:hover path:last-child {
-  fill: primaryAdd1;
-}
-
-.disabled {
-  composes: arrow;
-  pointer-events: none;
-}
 
 .contentTypeWrapper {
   width: 310px;
-}
-@media mobile {
-  .arrowSection {
-    display: block;
-  }
 }
