+++ packages/@coorpacademy-components/src/molecule/setup-sections/index.js
@@ -1,67 +1,9 @@
 import React, {useState, useCallback} from 'react';
 import PropTypes from 'prop-types';
-import classnames from 'classnames';
+import Draggable from '../draggable';
 import SetupSection from '../setup-section';
 import Loader from '../../atom/loader';
 import style from './style.css';
-
-const preventDefault = e => e.preventDefault();
-const stopPropagation = e => e.stopPropagation();
-
-const Draggable = props => {
-  const {id, dragging = false, onDragStart, onDragOver, onDragLeave, onDrop, children} = props;
-
-  const dragStartHandler = useCallback(
-    e => {
-      stopPropagation(e);
-      onDragStart(id);
-    },
-    [id, onDragStart]
-  );
-  const dragOverHandler = useCallback(
-    e => {
-      preventDefault(e);
-      onDragOver(id);
-    },
-    [id, onDragOver]
-  );
-  const dragLeaveHandler = useCallback(
-    e => {
-      preventDefault(e);
-      onDragLeave(id);
-    },
-    [id, onDragLeave]
-  );
-  const dropHandler = useCallback(
-    e => {
-      preventDefault(e);
-      onDrop(id);
-    },
-    [id, onDrop]
-  );
-
-  return (
-    <div
-      className={classnames(dragging ? style.dragging : null, style.section, style.draggable)}
-      onDragStart={dragStartHandler}
-      onDragOver={dragOverHandler}
-      onDragLeave={dragLeaveHandler}
-      onDrop={dropHandler}
-      draggable
-    >
-      {children}
-    </div>
-  );
-};
-Draggable.propTypes = {
-  id: PropTypes.string,
-  dragging: PropTypes.bool,
-  onDragStart: PropTypes.func,
-  onDragOver: PropTypes.func,
-  onDragLeave: PropTypes.func,
-  onDrop: PropTypes.func,
-  children: PropTypes.element
-};
 
 const Loading = () => (
   <div className={style.loading}>
@@ -107,7 +49,7 @@ const SetupSections = ({sections, loading, onDrop}) => {
 
   if (loading) return <Loading />;
 
-  const sectionsView = sections.map((section, index) => (
+  const sectionsView = sections.map(section => (
     <Draggable
       key={section.id}
       id={section.id}
@@ -117,12 +59,7 @@ const SetupSections = ({sections, loading, onDrop}) => {
       onDragLeave={dragLeaveHandler}
       onDrop={dropHandler}
     >
-      <SetupSection
-        {...section}
-        key={section.id}
-        onUp={index === 0 ? null : section.onUp}
-        onDown={index === sections.length - 1 ? null : section.onDown}
-      />
+      <SetupSection {...section} key={section.id} />
     </Draggable>
   ));
   return <div>{sectionsView}</div>;
