+++ packages/@coorpacademy-components/src/molecule/setup-sections/style.css
@@ -27,55 +27,3 @@
   min-height: 60px;
   position: relative;
 }
-.draggable {
-  cursor: grab;
-  position: relative;
-  background: transparent;
-}
-.draggable:hover{
-  background: rgba(0, 0, 0, 0.3);
-}
-.draggable:active{
-  background: transparent;
-}
-.draggable:before {
-  content: '.';
-  position: absolute;
-  left: 5px;
-  font-size: 20px;
-  line-height: 7px;
-  color: medium;
-  pointer-events: none;
-  text-shadow: 0 0 medium,    0 5px medium,    0 10px medium,    0 15px medium,    0 20px medium,    0 25px medium,    0 30px medium,    0 35px medium,    0 40px medium,    0 45px medium,
-             5px 0 medium,  5px 5px medium,  5px 10px medium,  5px 15px medium,  5px 20px medium,  5px 25px medium,  5px 30px medium,  5px 35px medium,  5px 40px medium,  5px 45px medium,
-            10px 0 medium, 10px 5px medium, 10px 10px medium, 10px 15px medium, 10px 20px medium, 10px 25px medium, 10px 30px medium, 10px 35px medium, 10px 40px medium, 10px 45px medium;
-}
-.dragging {
-  background: rgba(0, 0, 0, 0.1);
-}
-
-.dragging:after {
-  position: absolute;
-  content: "drop here";
-  left: 5px;
-  top: 5px;
-  bottom: 5px;
-  right: 5px;
-  background: color(black lightness(95%));
-  pointer-events: none;
-  border: 1px dashed medium;
-  box-sizing: border-box;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: dark;
-  font-family: 'Gilroy';
-  font-weight: 600;
-  font-size: 14px;
-}
-
-@media mobile {
-  .draggable:before {
-    content: none;
-  }
-}
